Have phpunit display the database connection / driver in use for tests
CI is currently always using the fallback database connection, and
therefore not running any of the platform/driver-specific tests.
Prior to fixing this, update the test utils to prove the issue by
displaying the details of the database connection in use the first
time it is loaded.

This debug information will also help developers be sure that they're
working with the expected environment when running tests locally.